fixes and code clean up

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -15,7 +16,10 @@ export class MessagesComponent implements OnInit {
     pageNumber = 1;
     pageSize = 5;
     loading = false;
-    constructor(private messageService: MessageService) { }
+    constructor(
+        private messageService: MessageService,
+        private confirmService: ConfirmService
+    ) {}
 
     ngOnInit(): void {
         this.loadMessages();
@@ -38,8 +42,17 @@ export class MessagesComponent implements OnInit {
     }
 
     deleteMessage(id: number) {
-        this.messageService.deleteMessage(id).subscribe(() => {
-            this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
-        });
+        this.confirmService
+            .confirm('Delete nude?', 'are u sure u want delete this message?')
+            .subscribe((result: boolean) => {
+                if (result) {
+                    this.messageService.deleteMessage(id).subscribe(() => {
+                        this.messages.splice(
+                            this.messages.findIndex((m) => m.id === id),
+                            1
+                        );
+                    });
+                }
+            });
     }
 }
